blog-app-backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
passlib
python-jose
python-dotenv
python-multipart
bcrypt